// KairosFrontend/src/app/services/model/eventDTO.ts
/**
 * Kairos Backend API
 * This API provides endpoints for the Kairos Backend application.
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface EventDTO { 
    id?: any;
    title: any;
    description: any;
    category: any;
    dateTime: any;
    maxParticipants: any;
    organizerId: any;
    structureId: any;
    sectors: any;
    images?: any;
}